+++ examples/next-openai/app/function-calling/page.tsx
@@ -63,10 +63,6 @@ export default function Chat() {
           ))
         : null}
       <div id="chart-goes-here"></div>
-      <div>
-        {' '}
-        Additional Data: <pre>{JSON.stringify(data)}</pre>
-      </div>
       <form onSubmit={handleSubmit}>
         <input
           className="fixed bottom-0 w-full max-w-md p-2 mb-8 border border-gray-300 rounded shadow-xl"