+++ examples/nuxt-openai/nuxt.config.ts
@@ -2,6 +2,9 @@
 export default defineNuxtConfig({
   devtools: { enabled: true },
   modules: ['@nuxtjs/tailwindcss'],
+  nitro: {
+    preset: 'vercel-edge' // you can use 'vercel' or other providers here
+  },
   runtimeConfig: {
     openaiApiKey: ''
   }