+++ examples/sveltekit-openai/src/routes/api/chat/+server.ts
@@ -1,4 +1,4 @@
-import { Configuration, OpenAIApi } from 'openai-edge'
+import OpenAI from 'openai'
 import { OpenAIStream, StreamingTextResponse } from 'ai'
 
 import { env } from '$env/dynamic/private'
@@ -9,17 +9,16 @@ import { env } from '$env/dynamic/private'
 import type { RequestHandler } from './$types'
 
 // Create an OpenAI API client
-const config = new Configuration({
-  apiKey: env.OPENAI_API_KEY
+const openai = new OpenAI({
+  apiKey: env.OPENAI_API_KEY || ''
 })
-const openai = new OpenAIApi(config)
 
 export const POST = (async ({ request }) => {
   // Extract the `prompt` from the body of the request
   const { messages } = await request.json()
 
   // Ask OpenAI for a streaming chat completion given the prompt
-  const response = await openai.createChatCompletion({
+  const response = await openai.chat.completions.create({
     model: 'gpt-3.5-turbo',
     stream: true,
     messages: messages.map((message: any) => ({